refactor: move http metrics layer to metrics mod
